Fix parameter case for create_guild_ban
Crystal doesn't support skewer-case-names, and this endpoint uses query string params with a normally body-based HTTP method, so we have to bypass the macro on this one and implement it manually.

I swear, this one endpoint is the bane of my existence.... It causes issues every time it shows up, without fail.